run dlrobot_central forever

+++ tools/robots/dlrobot/tests/cloud_complex/run.sh
@@ -9,27 +9,34 @@ RESULT_FOLDER=processed_projects
 WORKER_DIR=workdir
 SMART_PARSER_DB=smart_parser_cache.dbm
 rm -rf $RESULT_FOLDER *.log $INPUT_CENTRAl_FOLDER $INPUT_SMART_PARSER_FOLDER $SMART_PARSER_DB
-
 mkdir $INPUT_CENTRAl_FOLDER
 cp $PROJECT $INPUT_CENTRAl_FOLDER
-
-python3 ../../scripts/cloud/dlrobot_central.py --input-folder $INPUT_CENTRAl_FOLDER --result-folder  ${RESULT_FOLDER} &
-CENTRAL_PID=$!
-sleep 1
 
 python3 ../../scripts/cloud/smart_parser_cache.py --input-task-directory $INPUT_SMART_PARSER_FOLDER &
 SMART_PARSER_PID=$!
 sleep 1
 
-python3 ../../scripts/cloud/dlrobot_worker.py run_once --working-folder ${WORKER_DIR}
+python3 ../../scripts/cloud/dlrobot_central.py \
+    --input-folder $INPUT_CENTRAl_FOLDER \
+    --pdf-conversion-queue-limit 3000000000 \
+    --result-folder  ${RESULT_FOLDER} &
+CENTRAL_PID=$!
+sleep 2
 
-kill $SMART_PARSER_PID
-kill $CENTRAL_PID
+
+python3 ../../scripts/cloud/dlrobot_worker.py run_once --working-folder ${WORKER_DIR}
 
 sleep 3
 
-file_size=`wc -c $SMART_PARSER_DB | awk '{print $1}'`
-if [ $file_size -le 1000 ]; then
-  echo "broken smart parser json"
+
+python3 ../../scripts/cloud/smart_parser_cache_client.py --action stats > stats.json
+session_write_count=`cat stats.json | jq  .session_write_count`
+if [ "$session_write_count" != "1" ]; then
+  echo "bad session_write_count, must be 1"
+  kill $CENTRAL_PID
+  kill $SMART_PARSER_PID
+  exit 1
 fi
 
+kill $SMART_PARSER_PID
+kill $CENTRAL_PID
